* commit '0451730faea2642b759d22134a6f35dc394aa555':
  More 64-bit fixes for libcore native code.

+++ luni/src/main/java/libcore/icu/NativeDecimalFormat.java
@@ -621,7 +621,7 @@ public final class NativeDecimalFormat implements Cloneable {
     }
 
     private static native void applyPatternImpl(long addr, boolean localized, String pattern);
-    private static native int cloneImpl(long addr);
+    private static native long cloneImpl(long addr);
     private static native void close(long addr);
     private static native char[] formatLong(long addr, long value, FieldPositionIterator iter);
     private static native char[] formatDouble(long addr, double value, FieldPositionIterator iter);
